backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
azure-cognitiveservices-speech==1.42.0
python-dotenv==1.0.0